rag_agent/pyproject.toml
[project]
name = "rag-agent"
version = "0.1.0"
description = "RAG agent that answers questions about the project using the docs/knowledge_base corpus."
authors = [
    { name = "Google Hackathon Team", email = "hackathon@example.com" },
]
dependencies = [
    "google-adk>=1.18.0,<2.0.0",
    "google-cloud-aiplatform[evaluation,agent-engines]>=1.130.0",
    "google-cloud-storage>=2.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
]
requires-python = ">=3.10,<3.14"

[dependency-groups]
dev = [
    "pytest>=8.3.4,<9.0.0",
    "pytest-asyncio>=0.23.8,<1.0.0",
]

[project.optional-dependencies]
lint = [
    "ruff>=0.4.6,<1.0.0",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "C", "B", "UP", "RUF"]
ignore = ["E501", "C901", "B006"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = "."
asyncio_default_fixture_loop_scope = "function"

[tool.hatch.build.targets.wheel]
packages = ["app"]
